pyproject.toml
[tool.poetry]
name = "pipelines"
version = "1.0.0"
description = "Vertex AI Pipelines boilerplate."
authors = ["Carlos Ezequiel <cezequiel@google.com>"]
license = "Google"

[tool.poetry.dependencies]
python = "~3.10"
click = "8.1.3"
kfp = "1.8.13"
kubernetes = ">=8.0.0,<19"
cloudpathlib = "0.10.0"
PyYAML = ">=5.3,<6"
google-cloud-aiplatform = "1.17.0"
cryptography = "38.0.1"
rsa = "4.9"
sphinx_mdinclude = "^0.5.2"

[tool.poetry.dev-dependencies]
coverage = {extras = ["toml"], version = "6.4.4"}
pytest = "7.1.3"
pytest-cov = "3.0.0"
pytest-mock = "3.8.2"
black = "22.6.0"
flake8 = "5.0.4"
flake8-bandit = "4.1.1"
flake8-black = "0.3.3"
flake8-bugbear = "22.8.23"
flake8-import-order = "0.18.1"
safety = "2.1.1"
mypy = "0.971"
typeguard = "2.13.3"
flake8-annotations = "2.9.0"
flake8-docstrings = "1.6.0"
darglint = "1.8.1"
xdoctest = "1.0.1"
Sphinx = "5.1.1"
sphinx-autodoc-typehints = "1.19.0"
freezegun = "1.2.2"
types-PyYAML = "6.0.11"

[tool.poetry.scripts]
pipelines-cli = "pipelines.console:cli"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[build-system]
requires = ["poetry-core>=1.0.1"]
build-backend = "poetry.core.masonry.api"
